question.vueのリファクタリング
