+++ FRONTEND-ALCOHOL/src/stores/questions.ts
@@ -1,71 +1,42 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
 import axios from 'axios'
+import { useRouter } from 'vue-router'
 
 export const useQuestionsStore = defineStore('useQuestionsStore', () => {
-    type Genre = {
-        id: number
-        name: string
-        image: string
-        created_at: string
-        updated_at: string
-    }
 
-    type QuestionChoice = {
-        id: number
-        question_id: number
-        choice_text: string
-        question_number: number
-        score: number
-        created_at: string
-        updated_at: string
-    }
+    const questionNumber = ref(0)
+    const totalScore = ref(0)
+    const resultData = ref(null)
 
-    type Question = {
-        id: number
-        question_text: string
-        genre_id: number
-        created_at: string
-        updated_at: string
-        question_choices: QuestionChoice[]
-    }
-    const genres = ref<Genre[]>([])
-    const genre = ref<Genre>()
-    const questions = ref<Question[]>([])
-
-    const fetchAlcoholGenreMaster = async () => {
-        try {
-            const res = await axios.get('http://localhost/api/genres', {
-                withCredentials: true,
-                withXSRFToken: true,
-                headers: {
-                    Accept: 'application/json',
-                }
-            })
-            console.log('アルコールジャンル:', res.data)
-            genres.value = res.data;
-        } catch (error) {
-            console.error('アルコールジャンル取得失敗:', error)
+    const nextQuestion = async (id: number, score: number, genreId: number, questionsLength: number) => {
+        console.log(id + "問目", "スコア", score);
+        totalScore.value += score;
+        console.log("合計スコア", totalScore.value);
+        questionNumber.value++;
+        if (id >= questionsLength) {
+            // 最後の質問に到達した場合、結果画面に遷移
+            try {
+                const response = await axios.post('http://localhost/api/result', {
+                    genre_id: genreId,
+                    total_score: totalScore.value
+                }, {
+                    withCredentials: true,
+                    withXSRFToken: true,
+                    headers: {
+                        Accept: 'application/json',
+                    }
+                })
+                console.log('結果:', response.data)
+                resultData.value = response.data;
+                return true;
+            } catch (error) {
+                console.error('結果取得失敗:', error);
+                return false;
+            }
         }
+        return false;
     }
 
-    const fetchAlcoholGenreMasterDetail = async (genreId: number) => {
-        try {
-            const res = await axios.get(`http://localhost/api/genre/${genreId}`, {
-                withCredentials: true,
-                withXSRFToken: true,
-                headers: {
-                    Accept: 'application/json',
-                }
-            })
-            console.log('アルコールジャンル:', res.data.genre)
-            console.log('質問:', res.data.questions)
-            genre.value = res.data.genre;
-            questions.value = res.data.questions;
-        } catch (error) {
-            console.error('アルコールジャンル取得失敗:', error)
-        }
-    }
-
-    return { genres, fetchAlcoholGenreMaster, genre, questions, fetchAlcoholGenreMasterDetail }
+    return { nextQuestion, questionNumber, totalScore, resultData }
 })
