+++ FRONTEND-ALCOHOL/src/views/QuestionView.vue
@@ -1,12 +1,13 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
-import axios from 'axios'
 import { RouterLink, useRouter, useRoute } from 'vue-router'
 import { useAlcoholGenreMasterStore } from '@/stores/alcoholGenreMaster'
 import { useAuthUserStore } from '@/stores/authUser'
+import { useQuestionsStore } from '@/stores/questions'
 
 const alcohol_genre_master_store = useAlcoholGenreMasterStore()
 const auth_user_data_store = useAuthUserStore()
+const use_questions_store = useQuestionsStore()
 
 const router = useRouter();
 const route = useRoute();
@@ -19,38 +20,16 @@ onMounted(() => {
 	alcohol_genre_master_store.fetchAlcoholGenreMasterDetail(genreId)
 })
 
-const questionNumber = ref(0)
-const totalScore = ref(0)
-
 const logout = async () => {
 	await auth_user_data_store.logout()
 	router.push({ name: 'login' })
 }
 
-const choiceClick = async (id: number, score: number) => {
-	console.log("選択肢" , id, "スコア" , score);
-	totalScore.value += score;
-	console.log("合計スコア" , totalScore.value);
-	questionNumber.value++;
-	if (id >= alcohol_genre_master_store.questions.length) {
-		// 最後の質問に到達した場合、結果画面に遷移
-		try {
-			const response = await axios.post('http://localhost/api/result', {
-				genre_id: genreId,
-				total_score: totalScore.value
-			}, {
-				withCredentials: true,
-				withXSRFToken: true,
-				headers: {
-					Accept: 'application/json',
-				}
-			})
-			console.log('結果:', response.data)
-			router.push({ name: 'result'})
-		} catch (error) {
-			console.error('結果取得失敗:', error);
-		}
-	} 
+const choiceClickHandler = async (id: number, score: number) => {
+	const finishedFlag = await use_questions_store.nextQuestion(id, score, genreId, alcohol_genre_master_store.questions.length);
+	if (finishedFlag) {
+		router.push({ name: 'result'})
+	}
 }
 </script>
 
@@ -73,7 +52,7 @@ const choiceClick = async (id: number, score: number) => {
 		<section class="dark:bg-gray-100 dark:text-gray-800">
 			<div class="container p-4 mx-auto space-y-4 sm:p-6">
 				<div class="space-y-4">
-					<h3 class="text-2xl font-bold leading-none sm:text-5xl text-center">{{ alcohol_genre_master_store.questions[questionNumber]?.question_text }}</h3>
+					<h3 class="text-2xl font-bold leading-none sm:text-5xl text-center">{{ alcohol_genre_master_store.questions[use_questions_store.questionNumber]?.question_text }}</h3>
 				</div>
 				<div class="grid w-full grid-cols-1 gap-x-6 sm:grid-cols-2 lg:grid-cols-4 p-8">
 					<div class="space-y-4">
@@ -86,8 +65,8 @@ const choiceClick = async (id: number, score: number) => {
 		<section class="text-gray-600 body-font">
 			<div class="container px-5 py-4 mx-auto">
 				<div class="flex flex-wrap lg:w-4/5 sm:mx-auto sm:mb-2 -mx-2">
-					<div v-for="choice in alcohol_genre_master_store.questions[questionNumber]?.question_choices" :key="choice.id" class="p-2 sm:w-1/2 w-full">
-						<div class="bg-gray-100 rounded flex p-4 h-full items-center" @click="choiceClick(alcohol_genre_master_store.questions[questionNumber].id, choice.score)">
+					<div v-for="choice in alcohol_genre_master_store.questions[use_questions_store.questionNumber]?.question_choices" :key="choice.id" class="p-2 sm:w-1/2 w-full">
+						<div class="bg-gray-100 rounded flex p-4 h-full items-center" @click="choiceClickHandler(alcohol_genre_master_store.questions[use_questions_store.questionNumber].id, choice.score)">
 							<svg fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
 								stroke-width="3" class="text-indigo-500 w-6 h-6 flex-shrink-0 mr-4" viewBox="0 0 24 24">
 								<path d="M22 11.08V12a10 10 0 11-5.93-9.14"></path>
